+++ extensions/apple-reminders/src/hooks/useViewReminders.ts
@@ -136,7 +136,7 @@ export function groupByDueDates(reminders: Reminder[]) {
   }
 
   const remindersOnDate = useTimeOfDayGrouping
-    ? allDueDates.filter((date) => isBefore(date, today))
+    ? allDueDates.filter((date) => !isBefore(date, today) && date !== today)
     : allDueDates.filter((date) => date);
 
   remindersOnDate.forEach((date) => {
@@ -210,7 +210,7 @@ export default function useViewReminders(listId: string, { data }: { data?: Data
     { execute: showCompletedReminders },
   );
 
-  const viewDefault = listId === "today" || listId === "scheduled" ? "dueDate" : "default";
+  const viewDefault = listId === "today" || listId === "scheduled" || listId === "todo" ? "dueDate" : "default";
 
   const [sortBy, setSortBy] = useCachedState<SortByOption>(`sort-by-${listId}`, viewDefault);
   const [groupBy, setGroupBy] = useCachedState<GroupByOption>(`group-by-${listId}`, viewDefault);
@@ -221,6 +221,10 @@ export default function useViewReminders(listId: string, { data }: { data?: Data
       if (listId === "all") return true;
       if (listId === "today")
         return reminder.dueDate ? isOverdue(reminder.dueDate) || isToday(reminder.dueDate) : false;
+      if (listId === "todo")
+        return reminder.dueDate
+          ? isOverdue(reminder.dueDate) || (isFullDay(reminder.dueDate as string) && isToday(reminder.dueDate))
+          : false;
       if (listId === "scheduled") return !!reminder.dueDate;
       return reminder.list?.id === listId;
     };
@@ -298,6 +302,8 @@ export default function useViewReminders(listId: string, { data }: { data?: Data
           title = "All";
         } else if (listId === "today") {
           title = "Today";
+        } else if (listId === "todo") {
+          title = "Todo";
         } else if (listId === "scheduled") {
           title = "Scheduled";
         } else {